Finish testing QuoteService + update QuoteController

+++ springboot/src/main/java/ca/jrvs/apps/trading/domain/IexQuote.java
@@ -156,11 +156,11 @@ public class IexQuote {
   @JsonProperty("avgTotalVolume")
   private Integer avgTotalVolume;
   @JsonProperty("iexBidPrice")
-  private Integer iexBidPrice;
+  private Double iexBidPrice;
   @JsonProperty("iexBidSize")
   private Integer iexBidSize;
   @JsonProperty("iexAskPrice")
-  private Integer iexAskPrice;
+  private Double iexAskPrice;
   @JsonProperty("iexAskSize")
   private Integer iexAskSize;
   @JsonProperty("iexOpen")
@@ -591,12 +591,12 @@ public class IexQuote {
   }
 
   @JsonProperty("iexBidPrice")
-  public Integer getIexBidPrice() {
+  public Double getIexBidPrice() {
     return iexBidPrice;
   }
 
   @JsonProperty("iexBidPrice")
-  public void setIexBidPrice(Integer iexBidPrice) {
+  public void setIexBidPrice(Double iexBidPrice) {
     this.iexBidPrice = iexBidPrice;
   }
 
@@ -611,12 +611,12 @@ public class IexQuote {
   }
 
   @JsonProperty("iexAskPrice")
-  public Integer getIexAskPrice() {
+  public Double getIexAskPrice() {
     return iexAskPrice;
   }
 
   @JsonProperty("iexAskPrice")
-  public void setIexAskPrice(Integer iexAskPrice) {
+  public void setIexAskPrice(Double iexAskPrice) {
     this.iexAskPrice = iexAskPrice;
   }
